encoder_modification/training/requirements.txt
torch>=2.0
torchaudio>=2.0
numpy
scipy
matplotlib
soundfile
pyyaml
datasets
huggingface_hub
moshi
hydra-core 
omegaconf